[core] Move artifact extractor logic to classifier
