+++ conf/global.json
@@ -27,6 +27,11 @@
       "read_capacity": 5,
       "write_capacity": 5
     },
+    "artifact_extractor": {
+      "enabled": false,
+      "firehose_buffer_size": 128,
+      "firehose_buffer_interval": 900
+    },
     "firehose": {
       "use_prefix": true,
       "buffer_interval": 900,
