+++ streamalert/shared/__init__.py
@@ -1,7 +1,6 @@
 """Define some shared resources."""
 ALERT_MERGER_NAME = 'alert_merger'
 ALERT_PROCESSOR_NAME = 'alert_processor'
-ARTIFACT_EXTRACTOR_NAME = 'artifact_extractor'
 ATHENA_PARTITIONER_NAME = 'athena_partitioner'
 CLASSIFIER_FUNCTION_NAME = 'classifier'
 RULES_ENGINE_FUNCTION_NAME = 'rules_engine'
@@ -9,3 +8,5 @@ RULE_PROMOTION_NAME = 'rule_promotion'
 THREAT_INTEL_DOWNLOADER_NAME = 'threat_intel_downloader'
 
 CLUSTERED_FUNCTIONS = {CLASSIFIER_FUNCTION_NAME}
+
+ARTIFACTS_METRIC_NAME = 'artifacts'
