+++ streamalert/shared/metrics.py
@@ -18,7 +18,7 @@ import os
 from streamalert.shared import (
     ALERT_MERGER_NAME,
     ALERT_PROCESSOR_NAME,
-    ARTIFACT_EXTRACTOR_NAME,
+    ARTIFACTS_METRIC_NAME,
     ATHENA_PARTITIONER_NAME,
     CLASSIFIER_FUNCTION_NAME,
     RULES_ENGINE_FUNCTION_NAME
@@ -34,7 +34,6 @@ CLUSTER = os.environ.get('CLUSTER', 'unknown_cluster')
 # below when metrics are supported there
 FUNC_PREFIXES = {
     ALERT_MERGER_NAME: 'AlertMerger',
-    ARTIFACT_EXTRACTOR_NAME: 'ArtifactExtractor',
     CLASSIFIER_FUNCTION_NAME: 'Classifier',
     RULES_ENGINE_FUNCTION_NAME: 'RulesEngine'
 }
@@ -94,7 +93,7 @@ class MetricLogger:
             ALERT_ATTEMPTS: (_default_filter.format(ALERT_ATTEMPTS), _default_value_lookup)
         },
         ALERT_PROCESSOR_NAME: {},   # Placeholder for future alert processor metrics
-        ARTIFACT_EXTRACTOR_NAME: {
+        ARTIFACTS_METRIC_NAME: {
             EXTRACTED_ARTIFACTS: (_default_filter.format(EXTRACTED_ARTIFACTS),
                                   _default_value_lookup),
             FIREHOSE_FAILED_RECORDS: (_default_filter.format(FIREHOSE_FAILED_RECORDS),
