+++ streamalert_cli/_infrastructure/modules/tf_kinesis_firehose_setup/iam.tf
@@ -105,30 +105,3 @@ data "aws_iam_policy_document" "firehose_glue_catalog" {
     resources = ["*"]
   }
 }
-
-// IAM Policy: Invoke lambda function
-resource "aws_iam_role_policy" "streamalert_firehose_lambda" {
-  count = var.artifact_extractor_enabled ? 1 : 0
-  name  = "streamalert_firehose_invoke_lambda"
-  role  = "${aws_iam_role.streamalert_kinesis_firehose.id}"
-
-  policy = "${data.aws_iam_policy_document.firehose_lambda[0].json}"
-}
-
-// IAM Policy Document: Allow firehose to invoke artifact extractor lambda function
-data "aws_iam_policy_document" "firehose_lambda" {
-  count = var.artifact_extractor_enabled ? 1 : 0
-
-  statement {
-    effect = "Allow"
-
-    actions = [
-      "lambda:InvokeFunction",
-      "lambda:GetFunctionConfiguration"
-    ]
-
-    resources = [
-      "${var.function_alias_arn}*"
-    ]
-  }
-}
