+++ streamalert_cli/_infrastructure/modules/tf_kinesis_firehose_setup/variables.tf
@@ -26,15 +26,3 @@ variable "s3_logging_bucket" {
 variable "kms_key_id" {
   type = string
 }
-
-variable "artifact_extractor_enabled" {
-  type        = bool
-  default     = false
-  description = "Is Artifact Extractor Lambda function enabled"
-}
-
-variable "function_alias_arn" {
-  type        = string
-  default     = ""
-  description = "Fully qualified function arn of alias of Artifact extractor lambda"
-}
