+++ streamalert_cli/terraform/artifact_extractor.py
@@ -11,13 +11,12 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from streamalert.shared import ARTIFACT_EXTRACTOR_NAME
+# from streamalert.shared import ARTIFACT_EXTRACTOR_NAME
 from streamalert.shared.config import artifact_extractor_enabled, firehose_data_bucket
 from streamalert.shared.firehose import FirehoseClient
 from streamalert.shared.utils import get_database_name
 from streamalert_cli.athena.helpers import generate_artifacts_table_schema
 from streamalert_cli.terraform.common import infinitedict
-from streamalert_cli.terraform.lambda_module import generate_lambda
 
 # FIXME: Should we provide custom artifacs table name?
 DEFAULT_ARTIFACTS_TABLE_NAME = 'artifacts'
@@ -34,7 +33,7 @@ def generate_artifact_extractor(config):
     if not artifact_extractor_enabled(config):
         return
 
-    ae_config = config['lambda']['artifact_extractor_config']
+    ae_config = config['global']['infrastructure']['artifact_extractor']
     stream_name = FirehoseClient.artifacts_firehose_stream_name(config)
 
     # Set variables for the artifact extractor module
@@ -43,8 +42,6 @@ def generate_artifact_extractor(config):
         'account_id': config['global']['account']['aws_account_id'],
         'prefix': config['global']['account']['prefix'],
         'region': config['global']['account']['region'],
-        'function_role_id': '${module.artifact_extractor_lambda.role_id}',
-        'function_alias_arn': '${module.artifact_extractor_lambda.function_alias_arn}',
         'glue_catalog_db_name': get_database_name(config),
         'glue_catalog_table_name': ae_config.get('table_name', DEFAULT_ARTIFACTS_TABLE_NAME),
         's3_bucket_name': firehose_data_bucket(config),
@@ -54,17 +51,5 @@ def generate_artifact_extractor(config):
         'kms_key_arn': '${aws_kms_key.server_side_encryption.arn}',
         'schema': generate_artifacts_table_schema()
     }
-
-    # Set variables for the Lambda module
-    result['module']['artifact_extractor_lambda'] = generate_lambda(
-        '{}_streamalert_{}'.format(config['global']['account']['prefix'], ARTIFACT_EXTRACTOR_NAME),
-        'streamalert.artifact_extractor.main.handler',
-        ae_config,
-        config,
-        # Only pass Firehose stream name. Firehose client will translate it to full ARN
-        environment={
-            'DESTINATION_FIREHOSE_STREAM_NAME': stream_name
-        }
-    )
 
     return result
